Datenbank auf BigDecimal umgestellt, SHK Fortgesetzt, mit SyntaxDebugging begonnen. Vollständiges Semantisches Debugging fehlt noch

+++ src/excelchaos_model/database/Contract.java
@@ -3,6 +3,7 @@ package excelchaos_model.database;
 import com.j256.ormlite.field.DatabaseField;
 import com.j256.ormlite.table.DatabaseTable;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @DatabaseTable(tableName = "Contracts")
@@ -18,18 +19,18 @@ public class Contract {
     @DatabaseField()
     private Date end_date;
     @DatabaseField()
-    private double regular_cost;
+    private BigDecimal regular_cost;
     @DatabaseField()
-    private double bonus_cost;
+    private BigDecimal bonus_cost;
     @DatabaseField()
-    private double scope;
+    private BigDecimal scope;
     @DatabaseField()
     private String shk_hourly_rate;
     @DatabaseField()
     private boolean vbl_status;
 
-    public Contract(int id, String paygrade, String paylevel, Date start_date, Date end_date, double regular_cost,
-                    double bonus_cost, double scope, String shk_hourly_rate, boolean vbl_status) {
+    public Contract(int id, String paygrade, String paylevel, Date start_date, Date end_date, BigDecimal regular_cost,
+                    BigDecimal bonus_cost, BigDecimal scope, String shk_hourly_rate, boolean vbl_status) {
         this.id = id;
         this.paygrade = paygrade;
         this.paylevel = paylevel;
@@ -82,27 +83,27 @@ public class Contract {
         this.end_date = end_date;
     }
 
-    public double getRegular_cost() {
+    public BigDecimal getRegular_cost() {
         return this.regular_cost;
     }
 
-    public void setRegular_cost(double regular_cost) {
+    public void setRegular_cost(BigDecimal regular_cost) {
         this.regular_cost = regular_cost;
     }
 
-    public double getBonus_cost() {
+    public BigDecimal getBonus_cost() {
         return this.bonus_cost;
     }
 
-    public void setBonus_cost(double bonus_cost) {
+    public void setBonus_cost(BigDecimal bonus_cost) {
         this.bonus_cost = bonus_cost;
     }
 
-    public void setScope(double scope) {
+    public void setScope(BigDecimal scope) {
         this.scope = scope;
     }
 
-    public double getScope() {
+    public BigDecimal getScope() {
         return this.scope;
     }
