+++ src/excelchaos_model/database/SalaryIncreaseHistory.java
@@ -3,6 +3,7 @@ package excelchaos_model.database;
 import com.j256.ormlite.field.DatabaseField;
 import com.j256.ormlite.table.DatabaseTable;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @DatabaseTable(tableName = "SalaryIncreaseHistory")
@@ -10,7 +11,7 @@ public class SalaryIncreaseHistory {
     @DatabaseField()
     private int id;
     @DatabaseField()
-    private double new_salary;
+    private BigDecimal new_salary;
     @DatabaseField()
     private Date start_date;
     @DatabaseField()
@@ -18,7 +19,7 @@ public class SalaryIncreaseHistory {
     @DatabaseField()
     private boolean is_additional_payment;
 
-    public SalaryIncreaseHistory(int id, double new_salary, Date start_date, String comment, boolean is_additional_payment) {
+    public SalaryIncreaseHistory(int id, BigDecimal new_salary, Date start_date, String comment, boolean is_additional_payment) {
         this.id = id;
         this.new_salary = new_salary;
         this.start_date = start_date;
@@ -34,11 +35,11 @@ public class SalaryIncreaseHistory {
         return this.id;
     }
 
-    public double getNew_salary() {
+    public BigDecimal getNew_salary() {
         return this.new_salary;
     }
 
-    public void setNew_salary(double new_salary) {
+    public void setNew_salary(BigDecimal new_salary) {
         this.new_salary = new_salary;
     }
